ENORME

+++ ProjetAstronomy/src/ch/epfl/rigel/astronomy/Sun.java
@@ -26,7 +26,7 @@ public final class Sun extends CelestialObject{
      * @param meanAnomaly: mean anomaly of the sun.
      */
     public Sun(EclipticCoordinates eclipticPos, EquatorialCoordinates equatorialPos, float angularSize, float meanAnomaly) {
-        super("Soleil", equatorialPos, angularSize, (float)-26.7);
+        super("Soleil", equatorialPos, angularSize, -26.7f);
         Objects.requireNonNull(eclipticPos);
         this.eclipticPos = eclipticPos;
         this.meanAnomaly = meanAnomaly;
